requirements.txt
# taobao_user_behavior_analysis 项目依赖
# 创建时间：2026/2/5
# 更新：2026/2/9（锁定兼容版本）

# === 核心数据分析 ===
pandas==2.1.4
numpy==1.24.4

# === 数据格式和配置 ===
pyarrow==15.0.0
chardet==5.2.0
PyYAML==6.0.1

# === 可视化===
matplotlib==3.8.2
seaborn==0.13.0

# === 机器学习（可选）===
scikit-learn==1.3.2
xgboost==2.0.0
lightgbm==4.1.0

# === 开发工具（可选）===
jupyter==1.0.0
tqdm==4.66.1
python-dotenv==1.0.0

# === 代码质量（可选）===
black==23.12.1
pytest==7.4.4